# docs/swagger.yaml
definitions:
  global.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
      traceId:
        type: string
    type: object
...
host: 127.0.0.1:8080
info:
  contact:
    email: 25076778@qq.com
    name: API Support
    url: http://www.swagger.io/support
  description: Gin-base API 文档
  termsOfService: http://swagger.io/terms/
  title: Gin-base Swagger API
  version: "2.0"
paths:
  /api/v1/article:
    get:
      description: 文章列表
      parameters:
      - description: 认证Token
        in: header
        name: token
        required: true
        type: string
      - description: 页码
        in: query
        name: page
        required: true
        type: string
      - description: 分页大小
        in: query
        name: pageSize
        required: true
        type: string
      responses:
        "200":
          description: 成功返回" Example({"code":0,"msg":"Success","data":[]})
          schema:
            $ref: '#/definitions/global.Response'
        "400":
          description: 参数错误" Example({"code":400,"msg":"参数错误","data":[]})
          schema:
            $ref: '#/definitions/global.Response'
        "500":
          description: 系统错误" Example({"code":500,"msg":"系统错误","data":[]})
          schema:
            $ref: '#/definitions/global.Response'
      summary: 列表
      tags:
      - 文章管理
    post:
      consumes:
      - application/json
      description: 文章创建
      parameters:
      - description: 认证Token
        in: header
        name: token
        required: true
        type: string
      - description: 创建参数
        in: body
        name: data
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回" Example({"code":0,"msg":"Success","data":[]})
          schema:
            $ref: '#/definitions/global.Response'
        "400":
          description: 参数错误" Example({"code":400,"msg":"参数错误","data":[]})
          schema:
            $ref: '#/definitions/global.Response'
        "500":
          description: 系统错误" Example({"code":500,"msg":"系统错误","data":[]})
          schema:
            $ref: '#/definitions/global.Response'
      summary: 创建
      tags:
      - 文章管理
  /api/v1/article/{id}:
    delete:
      description: 文章删除
      parameters:
      - description: 认证Token
        in: header
        name: token
        required: true
        type: string
      - description: 文章ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: 成功返回" Example({"code":0,"msg":"Success","data":[]})
          schema:
            $ref: '#/definitions/global.Response'
        "400":
          description: 参数错误" Example({"code":400,"msg":"参数错误","data":[]})
          schema:
            $ref: '#/definitions/global.Response'
        "500":
          description: 系统错误" Example({"code":500,"msg":"系统错误","data":[]})
          schema:
            $ref: '#/definitions/global.Response'
      summary: 删除
      tags:
      - 文章管理
    get:
      description: 文章详情
      parameters:
      - description: 认证Token
        in: header
        name: token
        required: true
        type: string
      - description: 文章ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: 成功返回" Example({"code":0,"msg":"Success","data":[]})
          schema:
            $ref: '#/definitions/global.Response'
        "400":
          description: 参数错误" Example({"code":400,"msg":"参数错误","data":[]})
          schema:
            $ref: '#/definitions/global.Response'
        "500":
          description: 系统错误" Example({"code":500,"msg":"系统错误","data":[]})
          schema:
            $ref: '#/definitions/global.Response'
      summary: 详情
      tags:
      - 文章管理
    put:
      description: 文章更新
      parameters:
      - description: 认证Token
        in: header
        name: token
        required: true
        type: string
      - description: 文章ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新参数
        in: body
        name: data
        required: true
        schema:
          type: object
      responses:
        "200":
          description: 成功返回" Example({"code":0,"msg":"Success","data":[]})
          schema:
            $ref: '#/definitions/global.Response'
        "400":
          description: 参数错误" Example({"code":400,"msg":"参数错误","data":[]})
          schema:
            $ref: '#/definitions/global.Response'
        "500":
          description: 系统错误" Example({"code":500,"msg":"系统错误","data":[]})
          schema:
            $ref: '#/definitions/global.Response'
      summary: 更新
      tags:
      - 文章管理
  /api/v1/cache:
    delete:
      description: 删除缓存
      parameters:
      - description: 认证Token
        in: header
        name: token
        required: true
        type: string
      - description: 缓存键
        in: query
        name: key
        required: true
        type: string
      responses:
        "200":
          description: 成功返回" Example({"code":0,"msg":"Success","data":[]})
          schema:
            $ref: '#/definitions/global.Response'
        "400":
          description: 参数错误" Example({"code":400,"msg":"参数错误","data":[]})
          schema:
            $ref: '#/definitions/global.Response'
        "500":
          description: 系统错误" Example({"code":500,"msg":"系统错误","data":[]})
          schema:
            $ref: '#/definitions/global.Response'
      summary: 删除缓存
      tags:
      - 缓存管理
    get:
      description: 获取缓存
      parameters:
      - description: 认证Token
        in: header
        name: token
        required: true
        type: string
      - description: 缓存键
        in: query
        name: key
        required: true
        type: string
      responses:
        "200":
          description: 成功返回" Example({"code":0,"msg":"Success","data":[]})
          schema:
            $ref: '#/definitions/global.Response'
        "400":
          description: 参数错误" Example({"code":400,"msg":"参数错误","data":[]})
          schema:
            $ref: '#/definitions/global.Response'
        "500":
          description: 系统错误" Example({"code":500,"msg":"系统错误","data":[]})
          schema:
            $ref: '#/definitions/global.Response'
      summary: 获取缓存
      tags:
      - 缓存管理
    post:
      consumes:
      - application/json
      description: 设置缓存
      parameters:
      - description: 认证Token
        in: header
        name: token
        required: true
        type: string
      - description: 创建参数
        in: body
        name: data
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回" Example({"code":0,"msg":"Success","data":[]})
          schema:
            $ref: '#/definitions/global.Response'
        "400":
          description: 参数错误" Example({"code":400,"msg":"参数错误","data":[]})
          schema:
            $ref: '#/definitions/global.Response'
        "500":
          description: 系统错误" Example({"code":500,"msg":"系统错误","data":[]})
          schema:
            $ref: '#/definitions/global.Response'
      summary: 设置缓存
      tags:
      - 缓存管理
...
  /api/v1/send:
    post:
      description: http请求日志记录测试
      parameters:
      - description: 认证Token
        in: header
        name: token
        required: true
        type: string
      responses:
        "200":
          description: 成功返回" Example({"code":0,"msg":"Success","data":[]})
          schema:
            $ref: '#/definitions/global.Response'
        "500":
          description: 系统错误" Example({"code":500,"msg":"系统错误","data":[]})
          schema:
            $ref: '#/definitions/global.Response'
      summary: http请求日志记录测试
      tags:
      - 缓存管理
...
swagger: "2.0"